requirements.txt
Flask
tensorflow
numpy
Pillow
google-genai
python-dotenv
flask-cors